dev_d_one/api_d_one/requirements/requirements_backend.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
pytz==2022.7
sqlparse==0.4.3
gunicorn==20.1.0
whitenoise==6.4.0
djangorestframework-simplejwt